apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the Gossip semantic search engine"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.2",
    "httpx>=0.27.0",
    "linkup-sdk>=0.9.0",
    "openai>=2.8.1",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.16.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.14.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
